requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
PyYAML==6.0.2
python-dateutil==2.9.0.post0
jinja2==3.1.4
python-docx==1.1.2
reportlab==4.2.5
slowapi==0.1.9
playwright==1.48.0

# Testing & Code Quality (Qodo)
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
coverage==7.6.9
httpx==0.28.1
